Sketch to write to the ESP8266 (Aurduino)
ok, so I had to do this via the Aurduino interface, as I couldnt find a simple way to do so via Python.  Youll need the following:

https://www.arduino.cc/en/Main/Software    (Aurduino interface (I ran on Windows 10 no issues)
(once in the Aurduino shell, click File->Preferences.  In the "Additional Boards Manager URL text box, paste in the below:
https://arduino.esp8266.com/stable/package_esp8266com_index.json
Ok to close prefs.

At this point, theres 2 options..... one with WiFi support, one for using the 8266 as a micro controller stand alone.

Stand alone:


PLUG THE 8266 IN VIA USB (the com port will populate within aurduino interface in a second or 2).  set baud to 115200 (it will auto adjust faster).  Click Tools -> Board.  In the list of available boards, select the WEMOS D1 R1 and set the COM port 

Open the .ino sketch included in Aurduino.  Comment out (or leave in up to you) the wifi connect function.  Verify (should be good to go).  write (click the Right arrow in the menu area) to the 8266.  It will auto reboot the card.

At this point, the unit is up and running stand alone.

For WiFi, (if you cant python (3) in win..), open Python on your Pi (I used MU....)
Lead the Py file in.  Populate your SSID and Password... connect the 8266 to your Pi USB and run the script.  The script joins your stated SSID, and sets itself to AUTOCONNECT (even when power cycled).  Ive also married this to a IP reservation (DHCP on my rounter) to the cards MAC so it gets same IP every time.